r1108: Changing the permission on a file failed to update the display (reported by Karol Krenski). Also made other updates more efficient.

+++ ROX-Filer/src/find.c
@@ -112,7 +112,7 @@ struct _FindCondition
 	/* These next three depend on the first two... */
 	gpointer	data1;
 	gpointer	data2;
-	double		value;
+	gint		value;
 };
 
 struct _Eval
